requirements.txt
redis
quart
quart_discord
python-dotenv
jishaku
discord.py[speed]
asyncpg